# cmd/swagger/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "飞宏IM接口",
        "title": "飞宏IM接口",
        "contact": {
            "name": "飞宏IM技术支持",
            "url": "https://github.com/liony823/chat",
            "email": "ilovecoding@foxmail.com"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0.0"
    },
    "host": "localhost:10002",
    "basePath": "/",
    "paths": {
...
        "/statistics/user/register": {
            "post": {
                "description": "用户注册统计",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "statistics"
                ],
                "summary": "用户注册统计",
                "operationId": "userRegisterCount",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.UserRegisterCountReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.UserRegisterCountResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/get_users": {
            "post": {
                "description": "获取用户列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户",
                "operationId": "getUsers",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/user.GetPaginationUsersReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "$ref": "#/definitions/user.GetPaginationUsersResp"
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/get_users_online_status": {
            "post": {
                "description": "获取用户在线状态",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取用户在线状态",
                "operationId": "getUsersOnlineStatus",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/msggateway.GetUsersOnlineStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/msggateway.GetUsersOnlineStatusResp_SuccessResult"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        },
        "/user/get_users_online_token_detail": {
            "post": {
                "description": "获取在线用户详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "user"
                ],
                "summary": "获取在线用户详情",
                "operationId": "getUsersOnlineTokenDetail",
                "parameters": [
                    {
                        "description": "请求参数",
                        "name": "data",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/msggateway.GetUsersOnlineStatusReq"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "allOf": [
                                {
                                    "$ref": "#/definitions/apiresp.ApiResponse"
                                },
                                {
                                    "type": "object",
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "$ref": "#/definitions/msggateway.SingleDetail"
                                            }
                                        }
                                    }
                                }
                            ]
                        }
                    }
                }
            }
        }
    },
    "definitions": {
        "apiresp.ApiResponse": {
            "type": "object",
            "properties": {
                "data": {},
                "errCode": {
                    "type": "integer"
                },
                "errDlt": {
                    "type": "string"
                },
                "errMsg": {
                    "type": "string"
                }
            }
        },
...
        "msggateway.GetUsersOnlineStatusReq": {
            "type": "object",
            "properties": {
                "userIDs": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                }
            }
        },
        "msggateway.GetUsersOnlineStatusResp_SuccessDetail": {
            "type": "object",
            "properties": {
                "connID": {
                    "type": "string"
                },
                "isBackground": {
                    "type": "boolean"
                },
                "platformID": {
                    "type": "integer"
                },
                "token": {
                    "type": "string"
                }
            }
        },
        "msggateway.GetUsersOnlineStatusResp_SuccessResult": {
            "type": "object",
            "properties": {
                "detailPlatformStatus": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/msggateway.GetUsersOnlineStatusResp_SuccessDetail"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "msggateway.SingleDetail": {
            "type": "object",
            "properties": {
                "singlePlatformToken": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/msggateway.SinglePlatformToken"
                    }
                },
                "status": {
                    "type": "integer"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "msggateway.SinglePlatformToken": {
            "type": "object",
            "properties": {
                "platformID": {
                    "type": "integer"
                },
                "token": {
                    "type": "array",
                    "items": {
                        "type": "string"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        },
...
        "sdkws.RequestPagination": {
            "type": "object",
            "properties": {
                "pageNumber": {
                    "type": "integer"
                },
                "showNumber": {
                    "type": "integer"
                }
            }
        },
        "sdkws.UserInfo": {
            "type": "object",
            "properties": {
                "appMangerLevel": {
                    "type": "integer"
                },
                "createTime": {
                    "type": "integer"
                },
                "ex": {
                    "type": "string"
                },
                "faceURL": {
                    "type": "string"
                },
                "globalRecvMsgOpt": {
                    "type": "integer"
                },
                "nickname": {
                    "type": "string"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
...
        "user.GetPaginationUsersReq": {
            "type": "object",
            "properties": {
                "nickName": {
                    "type": "string"
                },
                "pagination": {
                    "$ref": "#/definitions/sdkws.RequestPagination"
                },
                "userID": {
                    "type": "string"
                }
            }
        },
        "user.GetPaginationUsersResp": {
            "type": "object",
            "properties": {
                "total": {
                    "type": "integer"
                },
                "users": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/sdkws.UserInfo"
                    }
                }
            }
        },
        "user.UserRegisterCountReq": {
            "type": "object",
            "properties": {
                "end": {
                    "type": "integer"
                },
                "start": {
                    "type": "integer"
                }
            }
        },
        "user.UserRegisterCountResp": {
            "type": "object",
            "properties": {
                "before": {
                    "type": "integer"
                },
                "count": {
                    "type": "object",
                    "additionalProperties": {
                        "type": "integer"
                    }
                },
                "total": {
                    "type": "integer"
                }
            }
        }
    }
}